requirements.txt
Fiona==1.7.0.post2
Flask==0.11.1
Flask-Celery-Helper==1.1.0
Flask-Redis==0.3.0
Jinja2==2.8
MarkupSafe==0.23
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
Shapely==1.5.16
Werkzeug==0.11.10
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
click==6.6
click-plugins==1.0.3
cligj==0.4.0
future==0.15.2
itsdangerous==0.24
kombu==3.0.35
modern-package-template==1.0
munch==2.0.4
oauthlib==1.1.2
python-twitter==3.1
pytz==2016.6.1
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
tweetmapper==0.1
wsgiref==0.1.2
